+++ R/quote.R
@@ -1,4 +1,4 @@
-# Copyright 2020 Cloudera Inc.
+# Copyright 2021 Cloudera Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,7 +16,7 @@
 NULL
 
 quote_column_in_expression <- function(expr, column) {
-  if (deparse(expr) == column)  {
+  if (deparse(expr) == column) {
     expr <- as.symbol(deparse(expr))
   }
   if (length(expr) == 1) {