+++ NEWS.md
@@ -1,3 +1,9 @@
+# tidyquery 0.2.3
+
+* Accepts names in form `namespace::object` in `FROM` clause ([#22](https://github.com/ianmcook/tidyquery/issues/22))
+* Exports its version number ([#23](https://github.com/ianmcook/tidyquery/issues/23))
+* Minor bugfixes and improvements
+
 # tidyquery 0.2.2
 
 * Improves stability with dtplyr and data.table ([#17](https://github.com/ianmcook/tidyquery/issues/17))